refactor/javaapi: refactor folders, refs: #43
