+++ src/main/java/client/api/BootApplication.java
@@ -1,4 +1,4 @@
-package ca.bc.gov.app;
+package client.api;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
