+++ src/main/java/client/api/core/entity/AbstractEntity.java
@@ -1,4 +1,4 @@
-package ca.bc.gov.app.core.entity;
+package client.api.core.entity;
 
 import java.io.Serializable;
