+++ src/main/java/client/api/core/repository/CoreRepository.java
@@ -1,4 +1,4 @@
-package ca.bc.gov.app.core.repository;
+package client.api.core.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
